chore: ImageInput 컴포넌트 내부 스타일링 로직 스타일 폴더로 분리
